EQMS-885 Refactor image and file extensions
